requirements.txt
openai
python-dotenv
requests
sentence_transformers
tiktoken
